feat(catalog): widen a recorded value to the full SH/SL pair
A SetVCP carries a 16-bit value and both backends already send both halves -
m1ddc's prepareDDCWrite splits a UInt16, and ddcutil's setvcp takes 0..65535 -
but the catalog stored one byte, so a reported value that needs the high byte
had nowhere to go. The 34U650A-B USB-C value, 465, was recorded in prose as
"a value that is not a byte" and left out of the table.

catalog.Operation, inputOp and the generator's Input.Value become uint16, and
each backend gains a private arguments(mechanism, value) builder that plan()
prepends the display selector to. Splitting it out lets a test pin the exact
argv for a value the catalog does not enable without any exported way to build
an Operation: rendering argv is not the same power as making a backend run one.

The 34U650A-B usb-c row goes in, disabled like the rest of that entry, and the
"a value that is not a byte" paragraph comes out. The tested model's argv is
byte-identical: values below 0x100 still render as two hex digits.

+++ internal/catalog/catalog.go
@@ -233,16 +233,20 @@ func (g Grade) String() string {
 // It is deliberately opaque. The fields are unexported and the constructor is
 // package-private, so the only way to obtain a valid Operation is to look one up
 // in this catalog. The zero value is invalid, and every backend rejects it.
+//
+// The value is 16 bits because DDC/CI carries one: a SetVCP writes an SH/SL
+// pair, and both backends already send both halves. Most recorded values fit in
+// a byte, and one does not.
 type Operation struct {
 	mechanism Mechanism
-	value     uint8
+	value     uint16
 	valid     bool
 }
 
 // newOperation builds a valid operation. It is package-private on purpose: no
 // caller outside the catalog may invent one. An unknown mechanism yields the
 // invalid zero value rather than something a backend might try to run.
-func newOperation(mechanism Mechanism, value uint8) Operation {
+func newOperation(mechanism Mechanism, value uint16) Operation {
 	if !mechanism.Known() {
 		return Operation{}
 	}
@@ -255,8 +259,9 @@ func (o Operation) Mechanism() Mechanism {
 	return o.mechanism
 }
 
-// Value returns the model-specific value to write.
-func (o Operation) Value() uint8 {
+// Value returns the model-specific value to write, as the 16-bit SH/SL pair a
+// SetVCP carries.
+func (o Operation) Value() uint16 {
 	return o.value
 }
 
@@ -296,7 +301,7 @@ func (i Identity) String() string {
 // merely reported elsewhere.
 type inputOp struct {
 	mechanism Mechanism
-	value     uint8
+	value     uint16
 	grade     Grade
 	evidence  string
 }
@@ -373,7 +378,7 @@ func (m Model) EnabledInputs() []Input {
 // is write-enabled. Documentation checks use it; nothing that writes does.
 //
 //nolint:gocritic // hugeParam: a value receiver keeps Model usable where it is not addressable
-func (m Model) InputValue(input Input) (uint8, bool) {
+func (m Model) InputValue(input Input) (uint16, bool) {
 	op, ok := m.Inputs[input]
 	if !ok {
 		return 0, false
